move snapshots to event sourcing directory
