+++ src/EventSourcing/SnapshotStore/sql_schema/event_store_snapshots.sql
