+++ tests/EventSourcingProviderTest.php
@@ -24,10 +24,10 @@ use Desperado\ServiceBus\EventSourcing\EventStreamStore\AggregateStore;
 use Desperado\ServiceBus\EventSourcing\EventStreamStore\Exceptions\NonUniqueStreamId;
 use Desperado\ServiceBus\EventSourcing\EventStreamStore\Sql\SqlEventStreamStore;
 use Desperado\ServiceBus\EventSourcingProvider;
-use Desperado\ServiceBus\EventSourcingSnapshots\SnapshotStore\SnapshotStore;
-use Desperado\ServiceBus\EventSourcingSnapshots\SnapshotStore\SqlSnapshotStore;
-use Desperado\ServiceBus\EventSourcingSnapshots\Snapshotter;
-use Desperado\ServiceBus\EventSourcingSnapshots\Trigger\SnapshotVersionTrigger;
+use Desperado\ServiceBus\EventSourcing\SnapshotStore\SnapshotStore;
+use Desperado\ServiceBus\EventSourcing\SnapshotStore\SqlSnapshotStore;
+use Desperado\ServiceBus\EventSourcing\Snapshotter;
+use Desperado\ServiceBus\EventSourcing\SnapshotTrigger\SnapshotVersionTrigger;
 use function Desperado\ServiceBus\Infrastructure\Storage\fetchOne;
 use Desperado\ServiceBus\Infrastructure\Storage\SQL\AmpPostgreSQL\AmpPostgreSQLAdapter;
 use Desperado\ServiceBus\Infrastructure\Storage\StorageAdapter;
